pyproject.toml
[project]
name = "valth-solutions-junior"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-ninja>=1.4.1",
    "httpx>=0.28.1",
    "load-dotenv>=0.1.0",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests"]

[tool.ruff]
exclude = ["migrations"]
